# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Up2You API Documentation",
        "description": "Documentazione delle API per la gestione di eventi e iscritti",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Recupera la lista di tutti gli eventi",
                "operationId": "7058ab7c89d20c752b6afd6d808f6526",
                "responses": {
                    "200": {
                        "description": "Lista degli eventi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Laravel Conference"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Una conferenza su Laravel."
                                            },
                                            "scheduled_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-03-30T10:00:00Z"
                                            },
                                            "location": {
                                                "type": "string",
                                                "example": "Milano"
                                            },
                                            "max_attendees": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Crea un nuovo evento",
                "operationId": "508895e03d7f10edac84357084c87b20",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "scheduled_at",
                                    "location",
                                    "max_attendees"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Laravel Conference"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Una conferenza su Laravel."
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-03-30T10:00:00Z"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Milano"
                                    },
                                    "max_attendees": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento creato con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Laravel Conference"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Una conferenza su Laravel."
                                        },
                                        "scheduled_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-30T10:00:00Z"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "Milano"
                                        },
                                        "max_attendees": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errore di validazione"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Recupera i dettagli di un evento",
                "operationId": "6f743f71f7a88dbe57be1d9c9ebf47cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dell'evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dettagli dell'evento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Laravel Conference"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Una conferenza su Laravel."
                                        },
                                        "scheduled_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-30T10:00:00Z"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "Milano"
                                        },
                                        "max_attendees": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento non trovato"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Aggiorna un evento esistente",
                "operationId": "fb130dcf1ab4d7abbd3a173741df1f8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dell'evento da aggiornare",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "scheduled_at",
                                    "location",
                                    "max_attendees"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Laravel Conference"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Una conferenza aggiornata su Laravel."
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-04-01T09:00:00Z"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Roma"
                                    },
                                    "max_attendees": {
                                        "type": "integer",
                                        "example": 150
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento aggiornato con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Laravel Conference"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Una conferenza aggiornata su Laravel."
                                        },
                                        "scheduled_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-04-01T09:00:00Z"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "Roma"
                                        },
                                        "max_attendees": {
                                            "type": "integer",
                                            "example": 150
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errore di validazione"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Elimina un evento",
                "operationId": "b8c8f4b41a32532f5e9f7146e511bc61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dell'evento da eliminare",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Evento eliminato con successo"
                    },
                    "404": {
                        "description": "Evento non trovato"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "API Key necessaria per accedere alle API",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}